refactor(core): extract legend definition logic
